+++ CHANGELOG.md
@@ -10,6 +10,9 @@ All notable changes to this project will be documented in this file. This projec
   - `isAny` to check if the value is one of any number of provided values.
   - `isEmpty` to check if the value is empty.
   - `isNotEmpty` to check if the value is not empty.
+  
+## Deprecated
+- Mutable value objects are deprecated and will be removed in 2.0
 
 ## [1.1.0] - 2017-07-25